refactor: extract common models & enums to reuse
