+++ src/Crowdin.Api/Reports/CostEstimationPostEditingGenerateReportRequest.cs
@@ -1,7 +1,6 @@
 ﻿
 using System;
 using System.Collections.Generic;
-using System.ComponentModel;
 
 using JetBrains.Annotations;
 using Newtonsoft.Json;
@@ -94,16 +93,6 @@ namespace Crowdin.Api.Reports
             [JsonProperty("taskId")]
             public int? TaskId { get; set; }
         }
-        
-        [PublicAPI]
-        public class BaseRatesForm
-        {
-            [JsonProperty("fullTranslation")]
-            public float FullTranslation { get; set; }
-                
-            [JsonProperty("proofread")]
-            public float Proofread { get; set; }
-        }
 
         [PublicAPI]
         public class IndividualRate
@@ -130,34 +119,6 @@ namespace Crowdin.Api.Reports
 #pragma warning disable CS8618
             public ICollection<Match> TmMatch { get; set; }
 #pragma warning restore CS8618
-        }
-
-        [PublicAPI]
-        public class Match
-        {
-            [JsonProperty("matchType")]
-            public MatchType MatchType { get; set; }
-            
-            [JsonProperty("price")]
-            public float Price { get; set; }
-        }
-        
-        [PublicAPI]
-        public enum MatchType
-        {
-            [Description("perfect")]
-            Perfect,
-
-            // ReSharper disable InconsistentNaming
-            [Description("100")]
-            Option_100,
-
-            [Description("99-82")]
-            Option_99_82,
-
-            [Description("81-60")]
-            Option_81_60
-            // ReSharper restore InconsistentNaming
         }
     }
 }
