+++ src/Crowdin.Api/Reports/GroupingParameter.cs
@@ -1,0 +1,16 @@
+﻿
+using System.ComponentModel;
+using JetBrains.Annotations;
+
+namespace Crowdin.Api.Reports
+{
+    [PublicAPI]
+    public enum GroupingParameter
+    {
+        [Description("user")]
+        User,
+                
+        [Description("language")]
+        Language
+    }
+}
